wip: documentation, rename init_state -> init_states, refactor solvers
