+++ jaxley/mechanisms/pumps/nernstreversal.py
@@ -48,6 +48,6 @@ class CaNernstReversal(Pump):
         """This dynamics model does not directly contribute to the membrane current."""
         return 0
 
-    def init_state(self, states, voltages, params, delta_t):
+    def init_states(self, states, voltages, params, delta_t):
         """Initialize the state at fixed point of gate dynamics."""
         return {}
