+++ tests/test_solver.py
@@ -8,7 +8,7 @@ import jaxley as jx
 from jaxley.connect import connect
 from jaxley.mechanisms.channels import HH
 from jaxley.mechanisms.synapses import IonotropicSynapse
-from jaxley.solver_gate import exponential_euler
+from jaxley.mechanisms.solvers import exponential_euler
 
 
 @pytest.mark.parametrize("x_inf", [3.0, 30.0])
